+++ ai/src/DataExtractor.kt
@@ -9,6 +9,7 @@ import klite.logger
 import klite.nodes.Node
 import klite.publicProperties
 import klite.warn
+import java.net.URI
 import kotlin.reflect.KClass
 import kotlin.reflect.KProperty1
 import kotlin.reflect.full.createType
@@ -20,14 +21,17 @@ class DataExtractor(
   private val log = logger()
   private val classPackageRegex = "\\b[\\w.]*\\.".toRegex()
 
-  fun <T: Any> extract(text: String, type: KClass<T>, provided: Map<KProperty1<T, *>, Any?> = emptyMap(), extraPrompt: String = "", numAttempts: Int = 3): T {
+  inline fun <reified T: Any> extract(text: String = "", imageUrl: URI? = null, provided: Map<KProperty1<T, *>, Any?> = emptyMap(), extraPrompt: String = ""): T =
+    extract(text, T::class, imageUrl, provided, extraPrompt)
+
+  fun <T: Any> extract(text: String, type: KClass<T>, imageUrl: URI? = null, provided: Map<KProperty1<T, *>, Any?> = emptyMap(), extraPrompt: String = "", numAttempts: Int = 3): T {
     val props = type.publicProperties - provided.keys.mapTo(mutableSetOf()) { it.name } - "id"
     val keys = props.values.joinToString { "${it.name}: " + it.returnType.toString().replace(classPackageRegex, "") }
     var prompt = "Output plain json with keys $keys, ISO dates, numbers as strings with dots: $text\n$extraPrompt"
     var response: AIClient.Response? = null
     repeat(numAttempts) {
       try {
-        response = aiClient.query(prompt, prevResponseId = response?.id)
+        response = aiClient.query(prompt, imageUrl = imageUrl, prevResponseId = response?.id)
         val jsonStr = response.text.stripMarkdown()
         if (provided.isEmpty()) return json.parse(jsonStr, type.createType())
         return type.createFrom(json.parse<Node>(jsonStr) + provided.mapKeys { it.key.name })